добавил в структуру userID
