+++ internal/shortener/handlers/util.go
@@ -1,0 +1,25 @@
+package handlers
+
+import (
+	"bytes"
+	"encoding/json"
+	"io"
+)
+
+func (h *APIHandler) unmarshalBody(body io.ReadCloser, form interface{}) error {
+
+	defer body.Close()
+
+	var buf bytes.Buffer
+	_, err := buf.ReadFrom(body)
+
+	if err != nil {
+		return err
+	}
+
+	if err = json.Unmarshal(buf.Bytes(), form); err != nil {
+		return err
+	}
+
+	return nil
+}
