+++ internal/shortener/request/request.go
@@ -1,0 +1,20 @@
+package request
+
+import (
+	"context"
+
+	"github.com/google/uuid"
+)
+
+type contextKey string
+
+const userIDKey = contextKey("UserID")
+
+func WithUserID(ctx context.Context, userID uuid.UUID) context.Context {
+	return context.WithValue(ctx, userIDKey, userID)
+}
+
+func UserIDFrom(ctx context.Context) (uuid.UUID, bool) {
+	v, ok := ctx.Value(userIDKey).(uuid.UUID)
+	return v, ok
+}
